forecast/requirements.txt
timesfm[torch]>=2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
huggingface-hub>=0.25.0
